+++ oscar/apps/catalogue/app.py
@@ -1,4 +1,4 @@
-from django.conf.urls.defaults import patterns, url, include
+from django.conf.urls import patterns, url, include
 
 from oscar.core.loading import get_classes
 from oscar.core.application import Application
@@ -19,19 +19,23 @@ class BaseCatalogueApplication(Application):
         urlpatterns = super(BaseCatalogueApplication, self).get_urls()
         urlpatterns += patterns('',
             url(r'^$', self.index_view.as_view(), name='index'),
-            url(r'^(?P<product_slug>[\w-]*)-(?P<pk>\d+)/$', self.detail_view.as_view(), name='detail'),
-            url(r'^(?P<category_slug>[\w-]+(/[\w-]+)*)/$', self.category_view.as_view(), name='category')
+            url(r'^(?P<product_slug>[\w-]*)_(?P<pk>\d+)/$',
+                self.detail_view.as_view(), name='detail'),
+            url(r'^(?P<category_slug>[\w-]+(/[\w-]+)*)/$',
+                self.category_view.as_view(), name='category')
         )
         return self.post_process_urls(urlpatterns)
 
 
 class ReviewsApplication(Application):
+    name = None
     reviews_app = reviews_app
 
     def get_urls(self):
         urlpatterns = super(ReviewsApplication, self).get_urls()
         urlpatterns += patterns('',
-            url(r'^(?P<product_slug>[\w-]*)-(?P<product_pk>\d+)/reviews/', include(self.reviews_app.urls)),
+            url(r'^(?P<product_slug>[\w-]*)-(?P<product_pk>\d+)/reviews/',
+                include(self.reviews_app.urls)),
         )
         return self.post_process_urls(urlpatterns)