+++ oscar/apps/catalogue/views.py
@@ -3,11 +3,14 @@ from django.views.generic import ListView, DetailView
 from django.db.models import get_model
 from django.utils.translation import ugettext_lazy as _
 
+from oscar.core.loading import get_class
 from oscar.apps.catalogue.signals import product_viewed, product_search
 
 Product = get_model('catalogue', 'product')
 ProductReview = get_model('reviews', 'ProductReview')
 Category = get_model('catalogue', 'category')
+ProductAlertForm = get_class('customer.forms',
+                             'ProductAlertForm')
 
 
 class ProductDetailView(DetailView):
@@ -15,32 +18,41 @@ class ProductDetailView(DetailView):
     model = Product
     view_signal = product_viewed
     template_folder = "catalogue"
-    _product = None
 
     def get_object(self):
-        if not self._product:
-            self._product = super(ProductDetailView, self).get_object()
+        # Use a cached version to prevent unnecessary DB calls
+        if not hasattr(self, '_product'):
+            setattr(self, '_product',
+                    super(ProductDetailView, self).get_object())
         return self._product
 
     def get_context_data(self, **kwargs):
         ctx = super(ProductDetailView, self).get_context_data(**kwargs)
-        ctx['reviews'] = ProductReview.objects.filter(status=ProductReview.APPROVED,
-                                                      product=self.object)
+        ctx['reviews'] = self.get_reviews()
+        ctx['alert_form'] = self.get_alert_form()
         return ctx
 
+    def get_alert_form(self):
+        return ProductAlertForm(user=self.request.user,
+                                product=self.object)
+
+    def get_reviews(self):
+        return self.object.reviews.filter(status=ProductReview.APPROVED)
+
     def get(self, request, **kwargs):
-        """
-        Ensure that the correct URL is used
-        """
+        # Ensure that the correct URL is used
         product = self.get_object()
         correct_path = product.get_absolute_url()
         if correct_path != request.path:
             return HttpResponsePermanentRedirect(correct_path)
-        response = super(ProductDetailView, self).get(request, **kwargs)
 
-        # Send signal to record the view of this product
-        self.view_signal.send(sender=self, product=product, user=request.user, request=request, response=response)
+        response = super(ProductDetailView, self).get(request, **kwargs)
+        self.send_signal(request, response, product)
         return response
+
+    def send_signal(self, request, response, product):
+        self.view_signal.send(sender=self, product=product, user=request.user,
+                              request=request, response=response)
 
     def get_template_names(self):
         """